+++ samples/server/petstore/jaxrs-cxf/src/gen/java/io/swagger/api/UserApi.java
@@ -14,6 +14,7 @@ import org.apache.cxf.jaxrs.ext.multipart.*;
 
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
+import javax.validation.constraints.*;
 
 @Path("/")
 @Api(value = "/", description = "")
@@ -53,7 +54,7 @@ public interface UserApi  {
     @Path("/user/login")
     @Produces({ "application/xml", "application/json" })
     @ApiOperation(value = "Logs user into the system", tags={ "user",  })
-    public String loginUser(@QueryParam("username")String username, @QueryParam("password")String password);
+    public String loginUser(@QueryParam("username") @NotNull String username, @QueryParam("password") @NotNull String password);
 
     @GET
     @Path("/user/logout")
