+++ samples/server/petstore/jaxrs-resteasy/default/pom.xml
@@ -108,10 +108,10 @@
             <version>2.7</version>
         </dependency>
         <dependency>
-              <groupId>io.swagger</groupId>
-                <artifactId>swagger-jaxrs</artifactId>
-                <version>${swagger-core-version}</version>
-              </dependency>
+            <groupId>io.swagger</groupId>
+            <artifactId>swagger-jaxrs</artifactId>
+            <version>${swagger-core-version}</version>
+        </dependency>
         <dependency>
             <groupId>junit</groupId>
             <artifactId>junit</artifactId>
@@ -138,6 +138,14 @@
                 </exclusion>
             </exclusions>
         </dependency>
+    <!-- Bean Validation API support -->
+    <dependency>
+        <groupId>javax.validation</groupId>
+        <artifactId>validation-api</artifactId>
+        <version>1.1.0.Final</version>
+        <scope>provided</scope>
+    </dependency>
+        
     </dependencies>
     <repositories>
         <repository>
