+++ samples/server/petstore/jaxrs-spec/src/gen/java/io/swagger/api/PetApi.java
