+++ samples/server/petstore/jaxrs-spec/src/gen/java/io/swagger/api/UserApi.java
