+++ samples/server/petstore/jaxrs-spec/src/gen/java/io/swagger/model/ModelApiResponse.java
