+++ samples/server/petstore/jaxrs-spec/src/gen/java/io/swagger/model/Order.java
