+++ samples/server/petstore/jaxrs-spec/src/gen/java/io/swagger/model/User.java
